Implemented loading card logic

+++ app/src/main/java/com/lifehackstudio/lifehacktest/home/CardAdapter.kt
@@ -38,7 +38,7 @@ class CardAdapter(private val listener: Listener) :
 
 
         holder.itemView.setOnClickListener {
-            listener.onItemClicked(cards[position].id)
+            listener.onItemClicked(cards[position].id.toLong())
         }
     }
